% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casetrim.R
\name{casetrim}
\alias{casetrim}
\title{Standardize letter case and trim whitespace for dataframes}
\usage{
casetrim(df, case = c("upper", "lower"), colclasses = c("asis",
  "character"))
}
\arguments{
\item{df}{a dataframe or list of dataframes with columns of any character type}

\item{case}{the letter case you want the output to be: upper or lower}

\item{colclasses}{do you want to return every column to it's original data type (asis) or simply return all cols as characters (character)?}
}
\description{
Removes leading and trailing whitespace from every column in a dataframe or list of dataframes. At the same time it standardizes the case of the characters to either upper or lower case. THis function works best with raw character data before converting data types and other data munging/feature engineering. Does handle a dataframe as it is, however, and returns every column in it's original data type as a dataframe.
}
\examples{
## Need examples...too lazy to make of find a dirty dataset to clean.
}
\keyword{case,}
\keyword{clean,}
\keyword{munge,}
\keyword{scrub}
\keyword{trim,}
