% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinecols.R
\name{combinecols}
\alias{combinecols}
\title{Combine two or more columns into one}
\usage{
combinecols(..., na.string = NA, returntype = c("character", "factor",
  "integer", "numeric"))
}
\arguments{
\item{...}{A vector of columns, a list of columns, or a data.frame that you want to collect into a single column.}

\item{na.string}{A character vector of string(s) which are to be interpreted as NA values. Blank observations are also considered to be missing values in logical, integer, numeric, and complex fields.}

\item{returntype}{The data type you want the combined column to be. Choose one of, 'character', 'factor', 'integer', 'numeric'.}
}
\description{
Takes a collection of columns as a list, a dataframe, or equal length vectors and merges them together. This is used when you want to combine 2 or more similar columns after a merge/join/rbind and the values don't overlap (if the values overlap, overlapped values will be lost). The collection can be various data types.
}
\examples{
df <- data.frame(performingPhysio = c("jill", "jack", '', ''), operatingPhysio = c(NA, NA, NA, "shankman"), stringsAsFactors = F)
df$physios <- combineCols(df$performingPhysio, df$operatingPhysio, na.string = c('', NA), returntype = 'character')
df

asd <- data.frame(col1 = as.factor(c(1:5, rep("", 15))), col2 = as.integer(c(rep("", 7), 8:13, rep("", 7))), col3 = c(rep("", 15), 16:20), stringsAsFactors = F)
qwe <- lapply(asd, function(x) x)

asd$newCol1 <- combineCols(asd$col1, asd$col2, asd$col3, na.string = '', returntype = 'integer')
asd$newCol2 <- combineCols(qwe, na.string = '', returntype = 'factor')
asd
}
\keyword{columns}
\keyword{combine}
