% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excelpmt.R
\name{excelpmt}
\alias{excelpmt}
\title{R analog/ rewrite to Excel's PMT function}
\usage{
excelpmt(rate, nper, pv, fv = 0, type = 0)
}
\arguments{
\item{rate}{interest rate}

\item{nper}{number of periods}

\item{pv}{present value}

\item{fv}{future value (default is 0)}

\item{type}{matureity of the payment; 0 = end of period, 1 = beginning; (default is 0)}
}
\description{
DESCRIPTION HERE
}
\examples{
EXAMPLES HERE
}
\keyword{amortization,}
\keyword{excel,}
\keyword{finance,}
\keyword{payment}
