% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notin.R
\name{\%!in\%}
\alias{\%!in\%}
\title{Not in (opposite to \%in\%)}
\usage{
x \%!in\% table
}
\arguments{
\item{x}{vector or NLL: the values to be matched. Long Vectors are suppored.}

\item{table}{vector or NULL: the values to be matched against. Long vectors are not supported.}
}
\description{
An easier and more logical way to express "elements of object X not in object Y"; \%in\% and \%!in\% are wrappers for match; returns a vector of booleans.
}
\examples{
1:10 \%!in\% 5:15
which(1:10 \%!in\% 5:15)
}
\keyword{in}
\keyword{match}
\keyword{not-in}
