% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left.R
\name{left}
\alias{left}
\title{R Analog to SQL's LEFT() function}
\usage{
left(vec, n, trimws = TRUE, sameclass = FALSE)
}
\arguments{
\item{vec}{A vector of any datatype. Not vectors of lists or matrices. I'm talking about charcter, numeric, logical, factor, etc.}

\item{n}{The number of characters you want to keep.}

\item{trimws}{Should the leading spaces be removed first? Default is \code{TRUE}.}

\item{sameclass}{Should the output be the same class as the input? Defaults to \code{FALSE}. Generally, you don't want this (expecially for POSIX classes which is impossible to do). Should the return vector be the same class as the input vector? Default is \code{FALSE} (returns a character vector no matter the input).}
}
\description{
Vectorized wrapper for taking the left N characters from a character, number, factor, etc. vector. Automatically trims leading whitespace. I find myself constantly trimming strings and grabbing only the first few characters so this saves some typing.
}
\examples{
left(vec = 'SomethingLong', n = 4)
left(vec = 40000.00, n = 2)
left(vec = '  AnotherThing', n = 4, trimws = FALSE)
left(vec = 400, 1, sameclass = TRUE)
}
\keyword{analog}
\keyword{left}
\keyword{sql}
\keyword{trimws}
