% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lettercases.R
\name{tocamel}
\alias{tocamel}
\alias{tocapital}
\title{Additional lettercase conversions}
\usage{
tocamel(x, collapseall = FALSE)

tocapital(x, fix_mc = FALSE, collapseall = FALSE)
}
\arguments{
\item{x}{A character vector, or list object of character vectors.}

\item{collapseall}{A logical scalar. Indicate whether or not you want a single character vector as your output. Default is \code{FALSE}.}

\item{fix_mc}{A logical scalar. Indicate if \code{Mc*} names \emph{(like McDowell)} should be double capitalized. Default is \code{FALSE}.}
}
\description{
Family of functions to change the lettercase of strings to camel or capital case.
}
\section{tocamel}{


Convert character vectors to \code{camelcase} \emph{(ie PaulJames)}

Takes a character vector, or list object of character vectors and converts the case to camelcase. Outputs the same object and number of elements as the input unless the collapseall flag is set to \code{TRUE}, in which case it will output a single character vector.
}

\section{tocapital}{


Convert character vectors to \code{capitalcase} \emph{(ie Paul James)}

Takes a character vector, or list object of character vectors and converts the case to capitalcase (AKA, title case). Outputs the same object and number of elements as the input unless the collapseall flag is set to \code{TRUE}, in which case it will output a single character vector.
}

\examples{
tocamel("jonny appleseed")
tocamel(c("jonny appleseed", "GraNDpa CrIPEs-mcgee"))

# using a list object
tocamel(
  list(
    c("THE", "DARK", "KNIGHT")
  , c("rises", "from the", "lazerus PIT")
  )
)

# using a list object and collapseall
tocamel(
  list(
    c("THE", "DARK", "KNIGHT")
  , c("rises", "from the", "lazerus PIT")
  )
  , collapseall = TRUE
)


tocapital("jonny appleseed")

# using a list object
tocapital(
  list(
    c("THE", "DARK", "KNIGHT")
  , c("rises", "from the", "lazerus PIT")
  )
)

# using a list object and collapseall
tocapital(
  list(
    c("THE", "DARK", "KNIGHT")
  , c("rises", "from the", "lazerus PIT")
  )
  , collapseall = TRUE
)

# examples of not fixing Mc* names (the default) vs fixing them
tocapital(c("jonny appleseed", "GraNDpa CrIPEs-mcgee"))
tocapital(
    c("jonny appleseed", "GraNDpa CrIPEs-mcgee")
  , fix_mc = TRUE
)
}
\seealso{
\code{\link{toupper}} \code{\link{tolower}}
}
\keyword{camel}
\keyword{capital}
\keyword{case}
\keyword{lettercase}
