% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mylib.R
\name{mylib}
\alias{mylib}
\title{Easily load multiple libraries without many library calls}
\usage{
mylib(..., suppress = TRUE)
}
\arguments{
\item{...}{An itemized listing of R package names to either install and load, or just load.}

\item{suppress}{A boolean indicator defaulted to TRUE that suppresses package messages when loaded. Set to \code{FALSE} to view all messages.}
}
\description{
Takes an itemized listing of R package names and loads them all. If the package is not currently installed, the function will attempt to install the package first. Currently only works if package is available on CRAN.
}
\examples{
\dontrun{
mylib(c("RPostgreSQL", "stringr", "dplyr"))
mylib("data.table")
}

}
\seealso{
\code{\link{library}}, \code{\link{require}}, \code{\link{install.packages}}
}
\keyword{install}
\keyword{library}
\keyword{package}
\keyword{require}
