% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notin.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Not in (opposite to \%in\%)}
\usage{
x \%nin\% table
}
\arguments{
\item{x}{vector or \code{NULL}: the values to be matched. Long Vectors are suppored.}

\item{table}{vector or \code{NULL}: the values to be matched against. Long vectors are not supported.}
}
\description{
An easier and more logical way to express "elements of object X not in object Y"; \%in\% and \%nin\% are wrappers for match; returns a vector of booleans.
}
\examples{
1:10 \%nin\% 5:15
which(1:10 \%nin\% 5:15)

}
\seealso{
\code{\link{\%in\%}}, \code{\link{match}}
}
\keyword{in}
\keyword{match}
\keyword{not-in}
