% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pft.R
\name{pft}
\alias{pft}
\title{Print the full time}
\usage{
pft(td = Sys.time())
}
\arguments{
\item{td}{A POSIX date time object or vector, or at least a string vector that can be interpretted by default as a POSIX object. Defaults to \code{Sys.time()} if no time is provided.}
}
\description{
Auto formats a POSIX object to output the day of week, full date, and 24 hour time.
}
\examples{
pft()

}
\seealso{
\code{\link{POSIXct}}, \code{\link{print}} \code{\link{pp}}
}
\keyword{POSIX}
\keyword{print}
\keyword{time}
