% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pjames.R
\docType{package}
\name{pjames}
\alias{pjames}
\alias{pjames-package}
\title{\code{pjames} package}
\description{
Paul's Personal Pacakge Panacea
}
\details{
See the README on:
\href{https://github.com/Paul-James/pjames#readme}{GitHub}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paul-james/pjames}
  \item Report bugs at \url{https://github.com/paul-james/pjames/issues}
}

}
\author{
\strong{Maintainer}: Paul James \email{paul.james379@gmail.com}

}
