% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp.R
\name{pp}
\alias{pp}
\title{Wrapper for paste and print}
\usage{
pp(...)
}
\arguments{
\item{...}{A collection of arguments for paste.}
}
\description{
Takes a vector of objects and puts them together in a vector of strings. It's useful to track progress of for loops and to automate output.
}
\examples{
pp(letters[1:10], 11:20)
for(i in 1:5){pp("This is the capital letter ", toupper(letters[i]), ".", sep = "")}
}
\keyword{paste}
\keyword{print,}
