% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlanalogs.R
\name{right}
\alias{right}
\alias{left}
\title{SQL Analog to left & right substring functions}
\usage{
right(vec, n, trimws = TRUE, sameclass = FALSE)

left(vec, n, trimws = TRUE, sameclass = FALSE)
}
\arguments{
\item{vec}{A vector of any datatype: \code{character, numeric, logical, factor}, etc.}

\item{n}{Integer scalar. The number of characters you want to keep.}

\item{trimws}{Should the applicable leading/trailing whitespace be removed first? Default is \code{TRUE}.}

\item{sameclass}{Should the output be the same class as the input? Defaults to \code{FALSE} \emph{(returns a character vector no matter the input)}. Generally, you don't want this \emph{(expecially for POSIX classes which is sketchy at best)}.}
}
\description{
Vectorized wrapper for taking the left/right N characters from a string, number, factor, etc., just like in SQL. I find myself constantly trimming strings and grabbing only the first/last few characters so this saves some typing.
}
\section{right}{


R Analog to SQL's \code{RIGHT()} function. Defaults to trim leading whitespace.
}

\section{left}{


R Analog to SQL's \code{LEFT()} function. Defaults to trim trailing whitespace.
}

\examples{
right(vec = 'SomethingLong', n = 4)
right(vec = 425575.4, n = 5)
right(vec = 'AnotherThing  ', n = 7, trimws = FALSE)
right(vec = 401.98, 4, sameclass = TRUE)

left(vec = 'SomethingLong', n = 4)
left(vec = 40000.00, n = 2)
left(vec = '  AnotherThing', n = 4, trimws = FALSE)
left(vec = 400, 1, sameclass = TRUE)
}
\keyword{analog}
\keyword{left}
\keyword{right}
\keyword{sql}
