% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tocamel.R
\name{tocamel}
\alias{tocamel}
\title{Convert character vectors to camelcase (ie PaulJames)}
\usage{
tocamel(x, collapseall = FALSE)
}
\arguments{
\item{x}{A character vector, or list object of character vectors.}

\item{collapseall}{A logical/boolean flag to indicate whether or not you want a single character vector as your output. Default is \code{FALSE}.}
}
\description{
Takes a character vector, or list object of character vectors and converts the case to camel case. Outputs the same object and number of elements as the input unless the collapseall flag is set to \code{TRUE}, in which case it will output a single character vector.
}
\examples{
tocamel("jonny appleseed")

# using a list object
tocamel(
  list(
    c("THE", "DARK", "KNIGHT")
  , c("rises", "from", "the", "lazerus", "PIT")
  )
)

# using a list object and collapseall
tocamel(
  list(
    c("THE", "DARK", "KNIGHT")
  , c("rises", "from", "the", "lazerus", "PIT")
  )
  , collapseall = TRUE
)

# examples of not fixing Mc* names (the default) vs fixing them
tocamel(c("jonny appleseed", "GraNDpa CrIPEs-mcgee"))
tocamel(
    c("jonny appleseed", "GraNDpa CrIPEs-mcgee")
  , fix_mc = TRUE
)

}
\seealso{
\code{\link{toupper}}, \code{\link{tolower}}, \code{\link{tocapital}}
}
\keyword{camelcase}
\keyword{capital}
\keyword{case}
