% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tocapital.R
\name{tocapital}
\alias{tocapital}
\title{Convert character vectors to capitalcase (ie Paul James)}
\usage{
tocapital(x, fix_mc = FALSE, collapseall = FALSE)
}
\arguments{
\item{x}{A character vector, or list object of character vectors.}

\item{fix_mc}{A logical/boolean flag to indicate if \code{Mc*} names (like McDowell) should be double capitalized. Default is \code{FALSE}.}

\item{collapseall}{A logical/boolean flag to indicate whether or not you want a single character vector as your output. Default is \code{FALSE}.}
}
\description{
Takes a character vector, or list object of character vectors and converts the case to capital case (AKA, title case). Outputs the same object and number of elements as the input unless the collapseall flag is set to \code{TRUE}, in which case it will output a single character vector.
}
\examples{
tocapital("jonny appleseed")

# using a list object
tocapital(
  list(
    c("THE", "DARK", "KNIGHT")
  , c("rises", "from", "the", "lazerus", "PIT")
  )
)

# using a list object and collapseall
tocapital(
  list(
    c("THE", "DARK", "KNIGHT")
  , c("rises", "from", "the", "lazerus", "PIT")
  )
  , collapseall = TRUE
)

# examples of not fixing Mc* names (the default) vs fixing them
tocapital(c("jonny appleseed", "GraNDpa CrIPEs-mcgee"))
tocapital(
    c("jonny appleseed", "GraNDpa CrIPEs-mcgee")
  , fix_mc = TRUE
)

}
\seealso{
\code{\link{toupper}}, \code{\link{tolower}}, \code{\link{tocamel}}
}
\keyword{camelcase}
\keyword{capital}
\keyword{case}
