% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tourney.R
\name{tourney}
\alias{tourney}
\title{Tournament seeding}
\usage{
tourney(...)
}
\arguments{
\item{...}{Each participant name as a character string separated by commas.}
}
\description{
Creates the seeding for a tournament bracket randomly given the names of the participants.
}
\examples{
tourney('Andrew', 'David', 'Keith', 'Paul', 'Jordan', 'Cristian', 'Matthew', 'Blair')

}
\keyword{bracket}
\keyword{seed}
\keyword{seeding}
\keyword{soccer-physics}
\keyword{tournament}
\keyword{tourney}
\keyword{wrestle-jump}
