% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibonacci.R
\name{fibonacci}
\alias{fibonacci}
\title{Calculate a fibonacci sequence}
\usage{
fibonacci(seq_length = 10, first_num = 0, second_num = 1)
}
\arguments{
\item{seq_length}{The length of the sequence as a number. The sequence length has to be greater than or equal to 2 (\code{seq_length >= 2}). The default is \code{10}.}

\item{first_num}{The starting point of the sequence. Use any positive number that comes before second_num (\code{first_num < second_num}). The default is \code{0}.}

\item{second_num}{The next number in the sequence after first_num. Use any positive number that comes after first_num (\code{second_num > first_num}). The default is \code{1}.}
}
\description{
Takes a sequence length and any two positive numbers in order to produce a fibonacci sequence. The fibonacci sequence is returned as a vector of numbers.
}
\details{
If using the default first and second numbers of \code{0} and \code{1} \R will return \code{Inf} starting at element number \code{1478}.
}
\examples{
\dontrun{
fibonacci('eight', 2, -1)  # wrong #1
fibonacci(8, 2, -1)  # wrong #2
fibonacci(8, 2, 1)  # wrong #3
fibonacci(8, 2, 2) # wrong #4
}
fibonacci(8, 10, 12)  # correct!
fibonacci(seq_length = 20, first_num = 10, second_num = 15)  # yay!
fibonacci(30)
fibonacci()

fibonacci(1500)[1477]
fibonacci(1500)[1478]
fibonacci(1500)[1500]

}
\keyword{fibonacci}
\keyword{math}
\keyword{nature}
\keyword{science}
\keyword{sequence}
