% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tocamel.R
\name{tocamel}
\alias{tocamel}
\title{Convert string to camelcase (ie PaulJames)}
\usage{
tocamel(..., sep = "\\\\W+", collapseall = FALSE)
}
\arguments{
\item{...}{A string, vector of strings, or list of strings in any letter case pattern.}

\item{sep}{A character vector containing characters or regular expression(s) to use for splitting. Default is "W+" (one or more characters that are not word characters; IE: not a letter or number). Uses perl regular expressions.}

\item{collapseall}{A logical/boolean flag to indicated whether or not you want a single string as your output (TRUE), or multiple strings (FALSE). Default is FALSE.}
}
\description{
Takes a string, vector(s) of strings, or list of strings and converts the case to camelcase. Outputs a string vector if only a single string vector was used as input. Outputs a list of string vectors if multiple string vectors were used as input. Outputs a single string if the collapseall flag is set to TRUE.
}
\examples{
tocamel("jonny appleseed")
tocamel(c("THE", "DARK", "KNIGHT"), c("rises from the lazerus PIT"))
tocamel(c("THE", "DARK", "KNIGHT"), c("rises from the lazerus PIT"), collapseall = TRUE)
tocamel("GraNDpa CrIPEs -mcgee!!!!")
}
\keyword{camel}
\keyword{camelcase}
\keyword{case}
