% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tocapital.R
\name{tocapital}
\alias{tocapital}
\title{Convert names to capitalcase (ie Paul James)}
\usage{
tocapital(..., sep = "\\\\s+", collapseall = F)
}
\arguments{
\item{...}{A string, vector of strings, or list of strings in any letter case pattern.}

\item{sep}{A character vector containing characters or regular expression(s) to use for splitting. Default is "\\s+" (one or more whitespace characters).}

\item{collapseall}{A logical/boolean flag to indicate whether or not you want a single string as your output (TRUE), or multiple strings (FALSE). Default is FALSE.}
}
\description{
Takes a string, vector(s) of strings, or list of strings and converts the case to capitalcase. Outputs a string vector if only a single string vector was used as input. Outputs a list of string vectors if multiple string vectors were used as input. Outputs a single string if the collapseall flag is set to TRUE.
}
\examples{
tocapital("jonny appleseed")
tocapital(c("THE", "DARK", "KNIGHT"), c("rises", "from", "the", "lazerus", "PIT"))
tocapital(c("THE", "DARK", "KNIGHT"), c("rises", "from", "the", "lazerus", "PIT"), collapseall = T)
tocapital("GraNDpa CrIPEs -mcgee")
}
\keyword{camelcase}
\keyword{capital,}
\keyword{case,}
